refactor: modularize clap stuff
