+++ src/types.rs
@@ -1,27 +1,4 @@
 pub mod chip;
 pub mod error;
 pub mod init_args;
-
-use clap::{Args, Parser, Subcommand};
-use init_args::InitArgs;
-
-#[derive(Parser)]
-#[command(name = "cargo")]
-#[command(bin_name = "cargo")]
-pub enum Cargo {
-    Embassy(Embassy),
-}
-
-#[derive(Args, Debug, Clone)]
-pub struct Embassy {
-    #[command(subcommand)]
-    pub command: EmbassyCommand,
-}
-
-#[derive(Debug, Clone, Subcommand)]
-pub enum EmbassyCommand {
-    #[command(about = "Initializes an Embassy project in the current workspace")]
-    Init(InitArgs),
-    #[command(about = "Opens the Embassy documentation page in your web browser")]
-    Docs,
-}
+pub mod parser;
