+++ src/types/parser.rs
@@ -1,0 +1,23 @@
+use crate::types::init_args::InitArgs;
+use clap::{Args, Parser, Subcommand};
+
+#[derive(Parser)]
+#[command(name = "cargo")]
+#[command(bin_name = "cargo")]
+pub enum Cargo {
+    Embassy(Embassy),
+}
+
+#[derive(Args, Debug, Clone)]
+pub struct Embassy {
+    #[command(subcommand)]
+    pub command: EmbassyCommand,
+}
+
+#[derive(Debug, Clone, Subcommand)]
+pub enum EmbassyCommand {
+    #[command(about = "Initializes an Embassy project in the current workspace")]
+    Init(InitArgs),
+    #[command(about = "Opens the Embassy documentation page in your web browser")]
+    Docs,
+}
